+++ packages/plugin/src/Bundles/Notifications/Providers/NotificationsProvider.php
@@ -27,7 +27,7 @@ class NotificationsProvider
 {
     public function __construct(private PropertyProvider $propertyProvider) {}
 
-    public function getById(int $id): ?NotificationInterface
+    public function getById(?int $id): ?NotificationInterface
     {
         $record = FormNotificationRecord::findOne(['id' => $id]);
         if (!$record) {